Organized `_Window(s)` class(es)
for future
