+++ main.py
@@ -1,16 +1,8 @@
 from PySide6 import QtWidgets
 
+from windows import BlessWindow
+
 import sys
-
-
-class BlessWindow(QtWidgets.QMainWindow):
-    def __init__(self):
-        super().__init__()
-        self.setWindowTitle("Bless")
-        button = QtWidgets.QPushButton("Mulai")
-
-        self.setCentralWidget(button)
-
 
 app = QtWidgets.QApplication(sys.argv)
